feat: enterprise UI upgrade — decompose scan page, responsive nav, graph DX
- Decompose scan/page.tsx (1,944 lines) into 4 modular components:
  scan-form.tsx, scan-result.tsx, attack-flow.tsx, scan-mesh.tsx
- Extract shared graph utilities (graph-utils.ts, graph-chrome.tsx)
  with DRY Controls/MiniMap/Background styling constants
- Add fullscreen toggle button to all 4 graph views
- Add responsive navigation with mobile hamburger menu and grouped
  nav sections (Scan, Inventory, Graphs, Govern)
- Upgrade stat cards with trend indicators and elevation shadows
- Improve severity chart with hover ring effects and drill-down hints
- Enhance graph edges with animated critical/high paths and refined
  marker sizing for the Turbo Flow pattern
- Proper dark mode elevation system (#0a0a0b base, layered shadows)
- Custom scrollbar styling and fullscreen CSS support
- Source breakdown cards with inner card elevation
